refactor: 메소드 라인 및 패키지 등 수정
