+++ src/main/java/baseball/app/ball/generator/BallsGenerator.java
@@ -2,6 +2,7 @@ package baseball.app.ball.generator;
 
 import baseball.app.ball.Balls;
 
+@FunctionalInterface
 public interface BallsGenerator {
 
     Balls generate();
