+++ src/main/java/baseball/app/constants/GameConstants.java
@@ -1,4 +1,4 @@
-package baseball.constants;
+package baseball.app.constants;
 
 public final class GameConstants {
