+++ src/main/java/baseball/app/game/Game.java
@@ -7,11 +7,12 @@ public class Game {
 
     private final Player player;
     private final Computer computer;
-    private boolean quit = false;
+    private boolean quit;
 
     public Game(final Player player, final Computer computer) {
         this.player = player;
         this.computer = computer;
+        this.quit = false;
     }
 
     public void start() {
@@ -19,12 +20,14 @@ public class Game {
             startGame();
             quit = quitGame();
         }
+
         GameView.printQuitGame();
     }
 
     private void startGame() {
         boolean isEnd = false;
         computer.prepareGame();
+
         while (!isEnd) {
             isEnd = playGame();
         }
@@ -36,8 +39,7 @@ public class Game {
     }
 
     private boolean quitGame() {
-        final GameMenu gameMenu = player.inputGameMenu();
-        return gameMenu == GameMenu.QUIT;
+        return player.inputGameMenu() == GameMenu.QUIT;
     }
 
 }
