+++ src/test/java/baseball/app/game/PlayerTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 
-import static baseball.constants.GameConstants.MAX_BALL_COUNT;
+import static baseball.app.constants.GameConstants.MAX_BALL_COUNT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
